+++ CHANGES.md
@@ -5,6 +5,7 @@
   decode non-UUID binary values
 - Zyp/Moksha/jq: `to_object` function now respects a `zap` option, that
   removes the element altogether if it's empty
+- Zyp/Moksha/jq: Improve error reporting at `MokshaTransformation.apply`
 
 ## 2024/09/22 v0.0.17
 - MongoDB: Fixed edge case when decoding MongoDB Extended JSON elements